refactor: create folder fetchers
