+++ front-end/src/lib/fetchers/octokitRequest.ts
@@ -2,9 +2,12 @@ import { Octokit, RequestError } from "octokit";
 import { Endpoints, RequestParameters } from "@octokit/types";
 import { toast } from "sonner";
 
-const octokit = new Octokit();
+export const octokit = new Octokit();
 
-export const octokitRequest = async <T>(route: keyof Endpoints, options: RequestParameters): Promise<T | undefined> => {
+export const octokitRequest = async <T>(
+  route: keyof Endpoints,
+  options: RequestParameters
+): Promise<T | undefined> => {
   try {
     const response = await octokit.request(route, options);
 
@@ -12,7 +15,9 @@ export const octokitRequest = async <T>(route: keyof Endpoints, options: Request
     const rateLimitRemaining = response.headers["x-ratelimit-remaining"];
 
     if (rateLimitRemaining && parseInt(rateLimitRemaining) < 10) {
-      console.warn(`GitHub API rate limit getting low: ${rateLimitRemaining} requests remaining`);
+      console.warn(
+        `GitHub API rate limit getting low: ${rateLimitRemaining} requests remaining`
+      );
     }
 
     return response.data as T;
@@ -22,10 +27,14 @@ export const octokitRequest = async <T>(route: keyof Endpoints, options: Request
       const rateLimitRemaining = headers["x-ratelimit-remaining"] === "0";
 
       if (error.status === 403 && rateLimitRemaining) {
-        const resetTime = Number(headers["x-ratelimit-reset"] || Date.now() / 1000);
+        const resetTime = Number(
+          headers["x-ratelimit-reset"] || Date.now() / 1000
+        );
         const resetDate = new Date(resetTime * 1000);
         const now = new Date();
-        const waitMinutes = Math.ceil((resetDate.getTime() - now.getTime()) / 60000);
+        const waitMinutes = Math.ceil(
+          (resetDate.getTime() - now.getTime()) / 60000
+        );
 
         const errorMessage = `Limite de requisições do GitHub atingido. Aguarde ${waitMinutes} minutos ou tente novamente após ${resetDate.toLocaleTimeString()}`;
         console.error(errorMessage);
@@ -35,7 +44,8 @@ export const octokitRequest = async <T>(route: keyof Endpoints, options: Request
       }
     }
 
-    const errorMessage = error instanceof Error ? error.message : "Erro desconhecido";
+    const errorMessage =
+      error instanceof Error ? error.message : "Erro desconhecido";
     console.error("Erro ao fazer requisição:", errorMessage);
     toast.error("Erro ao buscar repositórios. Tente novamente mais tarde.");
     return;
